+++ frontend/package.json
@@ -22,9 +22,9 @@
     "web-vitals": "^3.5.2"
   },
   "devDependencies": {
-    "@types/node": "^20",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
+    "@types/node": "^20.11.29",
+    "@types/react": "^18.2.67",
+    "@types/react-dom": "^18.2.22",
     "typescript": "^5"
   }
 }
