+++ frontend/src/components/CodeEditor.tsx
@@ -2,7 +2,7 @@ import { Box, HStack } from "@chakra-ui/react"
 import { Editor } from "@monaco-editor/react"
 import { useRef, useState } from "react"
 import LanguageSelector from "./LanguageSelector";
-import { CODE_SNIPPETS } from "../constants";
+import { CODE_SNIPPETS } from "../pages/constants";
 import Output from "./Output";
 
 const CodeEditor = () => {
